requirements.txt
lightning==2.0.9.post0
pytorch-lightning==2.0.9.post0
pytorch-toolbelt==0.6.3
torchmetrics==1.2.0
torchvision==0.16.0
pytorch-toolbelt==0.6.3